refactor(bench): extract shared harness machinery; signatures use real-cycle PMC
The cycle counter (just rewritten for bench-kem to read the core's real
CPU_CYCLES PMC via rdpmc, with rdtsc fallback + core pinning) was about to be
duplicated verbatim into bench/. Extract it — plus the median/stat helpers and
the per-op timing macro — into bench-common/harness_common.h, included by both
bench/harness.h and bench-kem/harness.h. Each harness now keeps only its
op-specific bench_run + output columns (keygen/sign/verify vs keygen/enc/dec).

This also brings the PMC counter to the signatures benchmark: bench/main.c prints
`# cyclecounter:` and bench/update_scheme_data.py records it in the env notes, so
both pages now measure real core cycles under the same methodology (cross-page
cycle comparison restored). Both build; bench-kem behaviour unchanged.
