+++ bench/update_scheme_data.py
@@ -263,10 +263,12 @@ def write_benchmark_env(meta: dict, path: Path) -> None:
     env["kernel"] = meta.get("kernel", "")
     env["compiler"] = meta.get("compiler", "")
     env["openssl"] = meta.get("openssl", "")
+    counter = meta.get("cyclecounter", "rdtsc + lfence (constant TSC rate, not execution cycles)")
     env["notes"] = (
         "Median over 1000 iterations (fewer for slow schemes). "
-        "Cycle counter: rdtsc + lfence (constant TSC rate, not execution cycles). "
-        "Wall clock: clock_gettime(CLOCK_MONOTONIC)."
+        f"Cycle counter: {counter}. "
+        "Wall clock: clock_gettime(CLOCK_MONOTONIC). "
+        "Benchmark thread pinned to one core."
     )
 
     sources = meta.get("_sources", {})
